+++ tools/IceRpc.Slice.Tools/IceRpc.Slice.Tools.csproj
@@ -56,7 +56,7 @@
       <PackagePath>tools/</PackagePath>
       <Pack>true</Pack>
     </None>
-    <None Include="../IceRpc.BuildTelemetry.Reporter/certs/">
+    <None Include="../IceRpc.BuildTelemetry.Reporter/certs/cacert.der">
       <PackagePath>tools/certs</PackagePath>
       <Pack>true</Pack>
     </None>
